resolve async cleanup
        - remove `struct Curl_dns_entry *dns` entry
        - remove `done` BIT
        - make port an uint16_t
        - make ip_version an uint8_t
        - rename `Curl_async_is_resolved()` to `Curl_async_take_result()`,
          returning CURLE_AGAIN while waiting on the result
        - remove `Curl_once_resolved()`

+++ lib/cf-socket.c
@@ -641,8 +641,8 @@ static CURLcode bindlocal(struct Curl_easy *data, struct connectdata *conn,
        * of the connection. The resolve functions should really be changed
        * to take a type parameter instead.
        */
-      int ip_version = (af == AF_INET) ?
-                       CURL_IPRESOLVE_V4 : CURL_IPRESOLVE_WHATEVER;
+      uint8_t ip_version = (af == AF_INET) ?
+                           CURL_IPRESOLVE_V4 : CURL_IPRESOLVE_WHATEVER;
 #ifdef USE_IPV6
       if(af == AF_INET6)
         ip_version = CURL_IPRESOLVE_V6;
